+++ NEWS.md
@@ -2,6 +2,7 @@
 ----------------
 
 - New defcustom, `org-d20-display-rolls-buffer`.
+- `org-d20-d20` shows the result of rolling a d4 in case /Bless/ is active.
 - Replace some calls to `loop` with calls to `cl-loop`.
 - Other code cleanup.